Update chart switch logic in Lens editor
