--- x-pack/plugins/lens/public/editor_frame_service/editor_frame/config_panel/layer_settings.test.tsx
@@ -1,63 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import React from 'react';
-import { screen } from '@testing-library/react';
-import faker from 'faker';
-import {
-  createMockDatasource,
-  createMockFramePublicAPI,
-  createMockVisualization,
-} from '../../../mocks';
-import { LayerSettings } from './layer_settings';
-import { renderWithReduxStore } from '../../../mocks';
-
-describe('LayerSettings', () => {
-  const renderLayerSettings = (propsOverrides = {}) => {
-    return renderWithReduxStore(
-      <LayerSettings
-        datasourceMap={{
-          testDatasource: createMockDatasource(),
-        }}
-        visualizationMap={{
-          testVis: createMockVisualization(),
-        }}
-        activeVisualization={createMockVisualization()}
-        layerConfigProps={{
-          layerId: 'myLayer',
-          state: {},
-          frame: createMockFramePublicAPI(),
-          setState: jest.fn(),
-          onChangeIndexPattern: jest.fn(),
-        }}
-        {...propsOverrides}
-      />
-    );
-  };
-
-  it('should render chart switch if custom layer header was not passed', () => {
-    renderLayerSettings({
-      activeVisualization: {
-        ...createMockVisualization(),
-        getDescription: () => ({ icon: 'myIcon', label: 'myVisualizationType' }),
-      },
-    });
-    expect(screen.getByTestId('lnsChartSwitchPopover')).toBeInTheDocument();
-  });
-
-  it('should use custom renderer if passed', () => {
-    const customLayerHeader = faker.lorem.word();
-
-    renderLayerSettings({
-      activeVisualization: {
-        ...createMockVisualization(),
-        getCustomLayerHeader: () => <div>{customLayerHeader}</div>,
-      },
-    });
-    expect(screen.getByText(customLayerHeader)).toBeInTheDocument();
-  });
-});
